% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{FizzBuzz Function}
\usage{
FizzBuzz(x)
}
\arguments{
\item{a}{A real number}

\item{b}{A real number}
}
\value{
FizzBuzz
}
\description{
Every time a number is divisible by 3, substitute "Fizz".
Every time a number is divisible by 5, substitute "Buzz".
Every time a number is divisible by both 3 and 5, substitute "FizzBuzz".
}
\examples{
FizzBuzz(1:30)
}
