% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{Weather recorded at Flagstaff Pulliam Airport 1950-2019.}
\format{
a data frame with 25449 observations
\describe{
 \item{DATE}{The date of observation.}
 \item{PRCP}{The amount of precipitation in inches}
 \item{SNOW}{The amount of snow in inches}
 \item{TMAX}{Max temperature in Fahrenheit}
 \item{Cold}{Min temperature in Fahrenheit}
}
}
\source{
\url{www.ncdc.noaa.gov/cdo-web/search}
}
\usage{
Flagstaff_Weather
}
\description{
Weather recorded at Flagstaff Pulliam Airport 1950-2019.
}
\keyword{datasets}
